moved all classes to their own files, as well as styles.
