+++ Styles.py
@@ -2,3 +2,5 @@
 TAN = "#DDC3A2"
 LIGHT = "#FFE4C4"
 FONT = "arial"
+BLACK = "#000000"
+WHITE = "#ffffff"
